Move message handling into its own module
With the test cases added, main.rs has become a bit messy, so split it out. Now
only main, docopt ardument definitions and crate/module imports are in it.
